Change to allow multiple webcams with same name.
The camera name appears to be based on the make and model. Having
two webcams attached of the same make and model means that one
"hides" the other. The fix is to add the device id into the
string that identifies the camera.

This could be improved by adding a map that would allow us to
use more friendly names like "cam1" "cam2" etc.